camera/requirements.txt
numpy==2.1.1
opencv-python==4.10.0.84
onvif-zeep==0.2.12